+++ example/lib/shared/decoration/potion_life.dart
@@ -2,7 +2,8 @@ import 'package:bonfire/bonfire.dart';
 import 'package:example/manual_map/dungeon_map.dart';
 import 'package:example/shared/util/common_sprite_sheet.dart';
 
-class PotionLife extends GameDecoration with Sensor<Player>, Movement {
+class PotionLife extends GameDecoration
+    with Sensor<Player>, Movement, BlockMovementCollision, BouncingObject {
   final double life;
   double _lifeDistributed = 0;
 
@@ -27,5 +28,11 @@ class PotionLife extends GameDecoration with Sensor<Player>, Movement {
     );
     removeFromParent();
     super.onContact(component);
+  }
+
+  @override
+  void onMount() {
+    moveUpLeft();
+    super.onMount();
   }
 }